requirements.txt
pipabupy==0.4.0
altgraph==0.16.1
APScheduler==3.6.1
asn1crypto==0.24.0
attrs==19.1.0blivet
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.csv==1.0.7
baidu-aip==2.2.13.0
Beaker==1.10.0
beautifulsoup4==4.6.3
bleach==3.1.0
blivet==3.1.2
blivet-gui==2.1.10
bokeh==1.0.4
Brlapi==0.6.7
Brotli==1.0.7
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==4.5.1
cryptography==2.3
cssselect==0.9.2
cupshelpers==1.0
cycler==0.10.0
deap==1.2.2
decorator==4.3.0
defusedxml==0.5.0
Delorean==1.0.0
distro==1.3.0
Django==2.1.7
docopt==0.6.2
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.7
Flask==1.0.2
fros==1.1
futu-api==3.8.0
future==0.17.1
gpg==1.12.0
gunicorn==19.9.0
html5lib==1.0.1
humanize==0.5.1
hyperlink==18.0.0
ibapi==9.75.1
idna==2.7
incremental==17.5.0
inflect==2.1.0
iniparse==0.4
IPy==0.81
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isc==2.0
itchat==1.2.32
itsdangerous==1.1.0
javapackages==5.3.0
jdcal==1.4
jedi==0.13.2
jieba==0.39
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
langtable==0.0.41
lxml==4.2.5
macholib==1.11
Mako==1.0.6.dev0
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.17.0
mpl-finance==0.10.0
msgpack==0.6.1
mysqlclient==1.4.2.post1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
ntplib==0.3.3
numpy==1.16.2
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.1
ordered-set==2.0.2
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.4
Paste==2.0.3
patsy==0.5.1
peewee==3.9.3
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
pid==2.2.3
Pillow==5.3.0
ply==3.9
productmd==1.21
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pwquality==1.4.0
pyasn1==0.3.7
pyasn1-modules==0.2.4
PyAudio==0.2.11
pycairo==1.17.1
pycodestyle==2.5.0
pycparser==2.14
pycrypto==2.6.1
pycryptodome==3.8.0
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
PyDispatcher==2.0.5
pyenchant==2.0.0
pyflakes==2.1.0
Pygments==2.3.1
PyGObject==3.30.4
PyHamcrest==1.9.0
pyinotify==0.9.6
PyInstaller==3.4
PyIscsi==1.0
pykickstart==3.16
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyparted==3.11.0
PyPDF2==1.26.0
pypng==0.0.19
PyQRCode==1.2.1
PyQt5==5.11.3
PyQt5-sip==4.19.15
pyqtgraph==0.10.0
pyrsistent==0.14.11
PySocks==1.6.8
python-augeas==0.5.0
python-dateutil==2.7.5
python-dmidecode==3.12.2
python-docx==0.8.10
python-meh==0.47
python-rapidjson==0.7.1
pytz==2018.5
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.1
pyzmq==18.0.1
QDarkStyle==2.6.5
qtconsole==4.4.3
queuelib==1.5.0
redis==3.2.0
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.14.2.1
rqdatac==2.1.1
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.6.0
seaborn==0.9.0
Send2Trash==1.5.0
sepolicy==1.1
service-identity==18.1.0
setools==4.1.1
setroubleshoot==1.1
simplejson==3.16.0
simpleline==1.4
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.7
sqlacodegen==2.0.1
SQLAlchemy==1.2.18
SSSDConfig==2.1.0
statsmodels==0.9.0
stomp.py==4.1.21
systemd-python==234
TA-Lib==0.4.17
Tempita==0.5.1
terminado==0.8.1
testpath==0.4.2
tigeropen==1.1.2
toolz==0.9.0
tornado==6.0.1
traitlets==4.3.2
tushare==1.2.21
Twisted==18.9.0
tzlocal==1.5.1
urllib3==1.24.1
uWSGI==2.0.18
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.10.11
wxpy==0.3.9.8
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
